% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FireCARES_1.R
\name{acs.dwnld}
\alias{acs.dwnld}
\title{Download ACS data}
\usage{
acs.dwnld(conn, year, cols, states = NULL)
}
\arguments{
\item{conn}{DBI Connection. The connection to the database where the data 
will be uploaded.}

\item{year}{Integer. The year for which to download ACS data}

\item{cols}{Character vector. Vector of column names (from the ACS) 
reflecting what ACS columns to download.}

\item{states}{Character vector. This allows me to specify a subset of states. 
If it is not specified, all nationwide data is downloaded. [optional]}
}
\value{
returns a list with the following entries:
  table.name.est: Name of the table on the database in which the new estimates
                  are stored. 
  table.name.err: Name of the table on the database in which the new error values
                  are stored.
  rows:           Number of rows added to the data set.
  elapsed.time:   Time it took to complete the download (in seconds?)
}
\description{
Downloads the ACS data needed for use in model estimation.
}
\details{
This leaves a 'temporary' table on the database. That table will then need
to be manually inserted into the main acs_est table.
}
