% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FireCARES_1.R
\name{c.test}
\alias{c.test}
\title{Merge multiple \code{\link{test}} objects}
\usage{
\method{c}{test}(t1, ...)
}
\arguments{
\item{t1}{test object.}

\item{...}{Additional test objects}
}
\value{
Returns a test object that contains all the information in the  
separate test objects supplied.
}
\description{
Merges multiple \code{\link{test}} objects
}
\examples{
dontrun{
  c.test(test.f.L1, test.f.L2)
}
}
