% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcTest.R
\name{c_test}
\alias{c_test}
\title{Merge multiple \code{\link{fcTest}} objects}
\usage{
c_test(t1, ...)
}
\arguments{
\item{t1}{test object.}

\item{...}{Additional test objects}
}
\value{
Returns a test object that contains all the information in the
separate test objects supplied.
}
\description{
Merges multiple \code{\link{fcTest}} objects
}
\examples{
\dontrun{
  c_test(test.f.L1, test.f.L2)
}
}
