% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{closeLog}
\alias{closeLog}
\title{Deletes an error log}
\usage{
closeLog()
}
\value{
This function invisibly returns the deleted error information.
}
\description{
This function deletes the error log.
}
\details{
This checks to see if an error logging environment has been set up,
and if it has, it deletes it.
}
