% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcCluster.R
\name{fcCluster}
\alias{fcCluster}
\title{Clusters the counties}
\usage{
fcCluster(conn, filter = 1000, clusters = 9)
}
\arguments{
\item{conn}{A DBI Connection. This is a connection to the database containing
the data and model definitions. If none is entered, default connection
information is obtained from the operating system environment.}

\item{filter=1000}{Numeric. If the number of missing fields in a column is greater
than \code{filter}, that column is excluded from the analysis. If
\code{filter} is NULL, then no columns are excluded.}

\item{clusters=9}{The number of clusters to use.}
}
\value{
Invisibly returns the revised cluster table used by later queries.
}
\description{
This function generates the county clusters for use by any estimation routine
that uses them.
}
\details{
The main purpose of this routine is to update the \code{nist.county_clusters} table
in the FireCARES database. It does the following tasks.

\itemize{
  \item Load needed data from the FireCARES database.
  \item Filter out the offshore territories and any columns with
  too few records to be effectively usable.
  \item Cluster the data using the cluster package
  \item Upload the county clustering results back to FireCARES.
}
}
