% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FireCARES_1.R
\name{fcMacro}
\alias{fcMacro}
\title{Run a set of control objects}
\usage{
fcMacro(npt, conn = NULL, save.tests = NULL)
}
\arguments{
\item{npt}{character vector. Lists names of the control objects to be run.}

\item{conn}{DBI connection. Connection to the database containing the 
'controls' sets. Only needed if the control objects in npt above are not
on the command line. [optional]}

\item{save.tests}{environment. Where to save the test results. [optional]}
}
\value{
nothing
}
\description{
This function takes a list of control objects and goes through the steps 
needed to run the tests for those control objects, and save the outputs.
}
\details{
The object listed in \code{npt} need not exist in the R environment. If it does not, then
this function calls the \code{\link{npt}} function to get the definition of the control object 
out of the database. That is what the 'conn' variable is used for. If the objects
already exist in the R environment, then there is no need to supply the \code{conn} variable.

If the save.tests environment is not supplied, then the test results will be removed 
from the R environment after they are saved to disk (and a summary saved to a data.frame). 
Some of these objects can be very large, so removing them is largely necessary. The \code{test} 
objects are the only ones saved. They are small enough that they can be saved without 
much harm. Note again that they are saved (if they are saved) in a separate environment 
which keeps the global environment less cluttered.

Creates rmse.sum in the global environment.
}
\examples{
dontrun{
  fcMacro(c("mr.final", "npt.final", "npt.final.L"))

  res <- new.env()
  fcMacro(c("mr.final", "npt.final", "npt.final.L"), save.tests=res)
}
}
