% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcRun.R
\name{fcRun}
\alias{fcRun}
\title{Fit models described in the supplied control objects}
\usage{
fcRun(sets, n = 0, err.Log = NULL)
}
\arguments{
\item{sets}{Control Object The control object describing the models to run.
This will typically be generated by 'npt'}

\item{n}{Integer. Number of bootstrap replications to run in order to
estimate the confidence intervals on parameters. n=0 (the default) will not
run any bootstrap replicates.}

\item{err.Log}{Either character or an err.Log structure. If this is a character
vector, then it creates an err.Log structure based on that vector. If it is
an err.Log structure, it uses the passed structure. If it is NULL, then no error
logging is done.}
}
\description{
The function takes the control object specifying a series of regression
models and runs those models.
}
\details{
Creates: an \code{out} object listing output of the models.

The reason that \code{out} is created in the global environment rather than
returned as an object is that if one of the models errors out, I still
get the results of the previous models.

This will typically be followed up by a run of \code{\link{fcTest}}
}
\examples{
\dontrun{
  fcRun(mr.d.00, sink="messages.13.txt")
  fcRun(mr.f.S0b, n=1000, sink="messages.08.txt")
  fcRun(mr.j.L0a)
}
}
