% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FireCARES_1.R
\name{fcTest}
\alias{fcTest}
\title{Compute out-of-sample RMS Errors for model output}
\usage{
fcTest(input, output, subset = NULL)
}
\arguments{
\item{input}{Control object. The input control object used by \code{\link{run}} to 
generate the output.}

\item{output}{Model Output. The model output produced by \code{\link{run}}.}

\item{subset}{The subset of the data over which to estimate RMS Errors. 
I include this because in some cases the test subset has been different from
the training subset in non-random ways.}
}
\description{
Compute out-of-sample RMS Errors for model output
}
\details{
This function takes output from the \code{\link{run}} function and calculates the
out-of-sample Root-Mean-Square Error values for each model in the output
object.
}
