% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_analysis.R
\name{full_analysis}
\alias{full_analysis}
\title{Runs the full analysis of a set of models.}
\usage{
full_analysis(conn = NULL, refresh.data = FALSE, models.run = NULL,
  merg = NULL, bypass.models = FALSE, do.predictions = TRUE,
  do.rmse = FALSE, roll.up.2.dept = TRUE, object.list = NULL,
  incl.detail = FALSE)
}
\arguments{
\item{conn}{A DBI Connection. This is a connection to the database containing
the data and model definitions. If none is entered, default connection
information is obtained from the operating system environment.}

\item{models.run}{Either a list or a data frame. This determines what models
are run. Its format is given as an example below. If it is
undefined, then a default set of models are run (see below).
Note that multiple model objects per risk level does not present
a problem.}

\item{merg}{Either a vector or a list. A value of '0' guarantees that the merg
option will not be used.}

\item{object.list}{List of data frames. The list must contain an entry for every risk
level that is run. The entry for each risk level must contain data frame
with the output from \code{\link{fcMacro}} for that risk level. If
\code{bypass.models} is TRUE and this parameter is undefined, the
function will error out. If \code{bypass.models} is FALSE, then
this parameter is ignored.}

\item{refresh.data=FALSE}{Logical. [This feature has not been implemented yet]
If this is TRUE, then the views on which the data for this
analysis are based are refreshed and the data.frames
used in this analysis are requeried.}

\item{bypass.models=FALSE}{Logical. If it is TRUE, then no models are estimated. If not,
then the models listed in 'models.run' above are estimated first.
Note that if 'bypass.models' is TRUE, then the 'objects'
data frame (output of the fcMacro function) must be supplied.}

\item{do.predictions=TRUE}{Logical. If it is TRUE, then predictions are generated from
the models run. If not, then no predictions are generated
from the estimated models.}

\item{do.rmse=FALSE}{Logical. If TRUE, then the RMS Errors of all the models will be
computed.}

\item{roll.up.2.dept=TRUE}{Logical. If it is TRUE, then the predictions are rolled
up to the department level. If it is FALSE, then the
predictions are left at the census tract level.}

\item{incl.detail=FALSE}{Logical. If TRUE, the \code{risk.results} section of the
output (described below) is included. If FALSE then the \code{risk.results}
section is not returned.}
}
\value{
returns a list with the following entries:

\describe{
\item{models.run}{The \code{models.run} input listing the models run by risk level.}

\item{bypass.models}{The input \code{bypass.models} value.}

\item{do.predictions}{The input \code{do.predictions} value.}

\item{roll.up.2.dept}{The input \code{roll.up.2.dept} value.}

\item{object.list}{The \code{object.list} object described above. If the
   \code{bypass.models} flag is set, then this is the object supplied to the
   function. Otherwise it is returned by the calls to \code{\link{fcMacro}}.}

\item{msg.name}{Name of the message file to which messages are sent.}

\item{prediction}{Data frame containing predictions for all variables requested
   in the models.run object. The predictions are either by census tract or by
   department depending on the value of the \code{roll.up.2.dept} flag.}

\item{risk.results}{This is a list, with an entry for each risk level. Each entry
   contains a data frame with the raw estimates for that risk level.
   For low and medium risk fires this contains the predictions at the
   census tract level (which are redundant with the results in
   \code{predictions} if \code{roll.up.2.dept} is FALSE). For high risk
   fires, this contains predictions at the parcel level.}
}
}
\description{
This function runs through the entire process of estimating models and
developing predictions at either the census tract level or at the department
level. This function can be run with NO inputs. That is, it can be called
as \code{full_analysis()} and it will work. Default values exist in some
form for all the parameters. Supply the parameters if you want different
results.
}
\details{
The \code{models.run} parameter can have one of two formats, a list format or a data
frame format. The list format is preferred. The data frame format has two columns:
\code{risk} and \code{lst}. Both columns have character format. Each row represents
a model set to be run. The \code{risk} column specfies the risk level associated with
that model set (and can only be one of 'lr', 'mr', 'hr', 'ems500', and 'emscty').
The \code{lst} column is the \code{lst} value from the \code{controls} database for the
model set to be run. The default value of \code{models.run} (in data.frame format) is
listed below.

\tabular{cc}{
\strong{risk} \tab \strong{lst}\cr
lr            \tab npt.final   \cr
mr            \tab mr.final    \cr
hr            \tab hr.final    \cr
ems500        \tab ems.5.final \cr
emscty        \tab ems.C.final
}

The list format contains a named entry for each risk level to be run. Each
entry contains a character vector listing the \code{lst} values from the
\code{controls} database for the model sets to be run for that risk level.
The default value of \code{models.run} (in list format) is listed below.

\code{models.run <- list(lr=c("npt.final"),
                         mr=c("mr.final"),
                         hr=c("hr.final"),
                         ems500=c("ems.5.final"),
                         emscty=c("ems.C.final"))}

The \code{merg} entry is if you want to merge multiple models into a
single model. It can either be a data frame or a list of data frames. The
data frame contains two columns. The first column, labeled either 'geoid' or
'tr10_fid', contains the tract/spatial identifiers used for prediction. The
'group' column should refer to a specific prediction column in the preliminary
output. See \code{\link{fcMerge}} for more details on format. If the
\code{merg} entry is a list, then each entry in the list is treated as
as separate merger and must have the appropriate format.

Each list entry can be (read 'should be') named. If it is named, then the
new column in the predictions output will have the name given the list entry.
If only a data frame is supplied, then the default name of "merge" will be used.
For that reason, the list format is preferred, even if only one merger is
done.

As currently written, this will only merge numeric columns. If you want to merge
non-numeric columns, the merger will have to be performed outside of this function.

This returns the output data frame with the new merged column and with the old
constituent columns removed. However, since the removal of the old columns is
performed last, it is possible to use the same source column for multiple final
merged columns.

The \code{merg} entry is only processed if \code{do.predictions} is
\code{TRUE}.

The \code{object.list} list object has an entry for each risk level run. That
entry is a data frame with information output from \code{\link{fcMacro}}. The
structure of that data frame is given by the following example:

\tabular{ccccc}{
\strong{npt.name} \tab \strong{res.name} \tab \strong{tst.name} \tab \strong{msg.name} \tab \strong{save.name} \cr
  npt.final       \tab npt.final.res     \tab npt.final.tst     \tab messages.00.txt   \tab npt.final.RData    \cr
  npt.final.L     \tab npt.final.L.res   \tab npt.final.L.tst   \tab messages.01.txt   \tab npt.final.L.RData
}

Note that if you are supplying the \code{object.list} structure while using the
\code{bypass.models} option, you can safely leave out the \code{tst.name} and
\code{msg.name} columns. Also, the \code{msg.name} column is now deprecated, so
do not rely on it.
}
