% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{getContext}
\alias{getContext}
\title{Retrieve message context.}
\usage{
getContext()
}
\value{
This function returns the message context in a character vector. If
logging has not been set up, or if no context has been set, NULL is returned.
}
\description{
This function returns the message context.
}
\details{
Message context is information that is printed with the message. The
intent is that message context will provide information that will help
contextualize the message.
}
