% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{getDests}
\alias{getDests}
\title{Returns the files to which messages are sent.}
\usage{
getDests(err.Log = NULL)
}
\arguments{
\item{err.Log}{the err.Log structure. If this is NULL, then the function returns
NULL}
}
\value{
This function returns a character vector containing the files to
which messages are sent. If an error environment has not been set
up, then it returns NULL.
}
\description{
Returns the files to which messages are sent.
}
\details{
This function returns the list of files to which messages are set.
It does not return 'console' in the vector, even if messages are
sent to it.
}
