% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcRun.R
\name{lasso}
\alias{lasso}
\title{LASSO helper function}
\usage{
lasso(formula, data, subset = NULL, ...)
}
\arguments{
\item{formula}{Formula. This describes the model that the LASSO fits.}

\item{data}{Data Frame. The data used for the model.}

\item{subset}{Name. This defines the subset of the data the model is
evaluated over.}

\item{...}{Additional parameters to the \code{\link[glmnet]{cv.glmnet}} function.}
}
\value{
Returns the glmnet.lasso object with the call slot altered to
reflect the call to this function rather than the glmnet function.
}
\description{
This is a helper function that \code{\link{fcRun}} calls whenever a LASSO model is used.
}
\details{
Basically it takes the standard inputs from the \code{\link{fcRun}} routine and translates
them to work with the glmnet \code{\link[glmnet]{cv.glmnet}} function.
}
