% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{msgOut}
\alias{msgOut}
\title{Send messages to the error log(s).}
\usage{
msgOut(message = NULL, type = "error")
}
\arguments{
\item{message}{character. The text of the message to be sent. [optional]}

\item{type}{character. The type of message. One of 'error', 'warning',
'message', 'status', 'start', or 'stop'. This parameter is not case
sensitive. Only the first option is used. Any unknown entry is treated
as equivalent to 'status'. See \code{details} for what these parameters
mean.}
}
\value{
This function invisibly returns the deleted error information.
}
\description{
This function sends messages to all error logs that have been set up.
}
\details{
The message that is posted depends on the type.

If \code{type} is one of 'error', 'warning', or 'message', then
the message (with \code{context} prepended) is posted to all logs.

If \code{type} is 'status' (or any equivalent--any unidentified type is
interpreted as 'status'), then the message (with \code{context} prepended)
is posted to the console IF console is one of the logs set up.

If \code{type} is 'start', then a message (with \code{context} prepended)
is posted to the console (IF console is one of the logs set up) stating
that a process has started. The 'start' code also records the starting
time.

If \code{type} is 'stop', then a message (with \code{context} prepended)
is posted to the console (IF console is one of the logs set up) stating
that a process has completed, with the elapsed time (if a starting time
has been recorded).

If no error logs have been set up, this routine automatically calls
\code{openLog} with a basic option to generate one log file.
}
