% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcTest.R
\name{naive}
\alias{naive}
\title{Naive estimator}
\usage{
naive(test)
}
\arguments{
\item{test}{\code{\link{fcTest}} object.}
}
\value{
(modified) \code{\link{fcTest}} object.
}
\description{
Generates the naive estimator for any given data set.
}
\details{
This takes an output object from the \code{\link{fcTest}} function and computes
the Naive predictor. The Naive predictor says that the best prediction
for a tract-year is the number of outcomes (fires, injuries, etc.)
that occurred for that tract the previous year. It is undefined
for the first year in the data set.
}
