% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{openLog}
\alias{openLog}
\title{Open an error log}
\usage{
openLog(dest = c(NA, "console"))
}
\arguments{
\item{dest}{character vector. destinations to send the error output to.
Destinations are usually file names.}
}
\value{
This function invisibly returns the stored error information.
}
\description{
This function opens the error log.
}
\details{
This sets up the error logging system. The parameter \code{dest} lists
all the output locations to which messages are to be sent. If more
than one location is specified, then the messages are sent to all
the specified locations.

With three exceptions, the values in \code{dest} are interpreted
as file names. The exceptions are 'console' (case is ignored here),
the empty string (''), and NA. Either of the first two cases are
interpreted as requesting that output be sent to the console.

For every NA included in the \code{dest} vector, a file name
is generated. The file name has the form 'messages.xx.txt' where
the 'xx' is a two (or more) digit number, starting with '01'.
The function finds the largest such file name in the working directory
and uses the next larger number for the new file name. So, I can
auto-generate 3 (or any other arbitrary number) of log files simply
by listing \code{NA} three times in the \code{dest} parameter vector.

Duplicates are automatically removed. Note that (aside from the
'console'), names are case sensitive. So two entries with the same
name, aside from case differences, will both be retained. For operating
systems that are case insensitive, that will result in duplicate
messages being sent to the file.

If the session is not interactive, then the request to send messages
to the console is automatically ignored.

If an error environment has been set up for a calling function, then
some of the information for the calling function is imported for this
function. In particular, \code{context} is imported, and the \code{dest}
values are imported. If there is an error environment for a calling
function, then no new file name is generated, even if \code{NA} is
included in this call. As usual, duplicates are removed. So, any
messages are sent to all log files used for calling functions, and
to any new files specifically specified in this call. But no new
file is automatically generated, even if \code{NA} is specified in
this call.

If an error environment has already been set up for this
function, this deletes it and starts over.
}
