% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FireCARES_1.R
\name{ranger}
\alias{ranger}
\title{ranger helper function}
\usage{
ranger(formula, data, subset = NULL, ...)
}
\arguments{
\item{formula}{Formula. This describes the model that the LASSO fits.}

\item{data}{Data Frame. The data used for the model.}

\item{subset}{Name. This defines the subset of the data the model is 
evaluated over.}

\item{...}{Additional parameters to the \code{\link[ranger]{ranger}} function.}
}
\value{
Returns the ranger object with the call slot altered to 
reflect the call to this function rather than the ranger function.
}
\description{
This is a helper function that \code{\link{run}} calls whenever a ranger model is used.
}
\details{
Basically it takes the standard inputs from the 'run' routine and translates
them to work with the \code{link[ranger]{ranger}} function.

The reason this helper function exists is because the default ranger
function does not have a subset argument.
}
