% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollUp2Dept.R
\name{rollUp2Dept}
\alias{rollUp2Dept}
\title{Roll up census-tract predictions to the department level}
\usage{
rollUp2Dept(predictions, fire.col, sz2.col, sz3.col, indx.cols)
}
\arguments{
\item{predictions}{data frame containing predictions as output by the fcEstimate function. See details below.}

\item{fire.col}{character vector containing names of the columns with fire predictions.}

\item{sz2.col}{character vector containing names of the columns with predicted percentages of fires that
go beyond the room of origin.}

\item{sz3.col}{character vector containing names of the columns with predicted percentages of size 2 fires
that go beyond the structure of origin.}

\item{indx.cols}{character vector containing the names of the columns to which the data is
to be rolled up.}
}
\value{
A data frame with the department-level predictions.
}
\description{
This function takes output from the fcEstimate function, which is typically at the census-tract
level, and rolls them up to the department level.
}
\details{
This routine is intended to work with a data frame containing multiple estimates (think low- medium-
and high-risk estimates all contained in the same data frame). When that is the case, the first
entry in fire.col goes with the first entry in sz2.col and sz3.col. The second entry in in fire.col
goes with the second entry in sz2.col and sz3.col. And so on. See the example below.

This function takes output from the fcEstimate function, which is typically at the census-tract
level, and rolls them up to the department level.
What makes this more complicated than a simple call to aggregate is that the sz2 and sz3
columns are percentages rather than estimated counts. So, summing those columns produces
nonsensical results.

What this function does is create temporary columns for the estimated counts for medium and
large fires. It then sums over census tracts (rolling up to the department level), and
back-calcuates the percentages from the accumulated estimated counts.
}
\examples{
\dontrun{
dept.predict <- rollUp2Dept(predictions, c("lr.fires", "mr.fires", "hr.fires"),
                                         c("lr.sz2",   "mr.sz2",   "hr.sz2"),
                                         c("lr.sz3",   "mr.sz3",   "hr.sz3"),
                                         c("year", "fd_id"))
}
}
