% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{setContext}
\alias{setContext}
\title{Sets message context.}
\usage{
setContext(err.Log = NULL, context)
}
\arguments{
\item{err.Log}{The err.log structure. If this is NULL, then the function
does nothing.}

\item{context}{Character Vector. A character vector representing the context
information to be printed with the messages.}
}
\value{
This function returns the updated err.Log structure.
}
\description{
This function sets the message context.
}
\details{
Message context is information that is printed with the message. The
intent is that message context will provide information that will help
contextualize the message.

If no logging has been set up, this automatically calls the setup routine
with the type set to generate one message file.
}
