% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcSetup.R
\name{fcSetup}
\alias{fcSetup}
\title{Condition a data set for use in model estimation}
\usage{
fcSetup(dta, seed = 953016876)
}
\arguments{
\item{dta}{data.frame. The data set that needs to be condition for use.}

\item{seed}{Intger. A random seed used to ensure consistent results
for the partitioning of the data set into training and test sets.}
}
\value{
The conditioned data frame.
}
\description{
Condition a data set for use in model estimation
}
\details{
This takes the 'low.risk.fires', 'med.risk.fires' and 'high.risk.fires'
data.frames as pulled from the database and makes the modification needed
to use them for analysis.

It also does much of the pre-processing on the 'lr_mr_pred' and 'hr_pred'
data frames as well.

This function carries out the following tasks:
\enumerate{
  \item Set any nulls in the outcome variables to zero.
  \item Turn any categorical predictors into factors.
  \item Take the log of the income variable.
  \item Ensure that there is an f_located column in the table.
  \item Define any filters that are needed (only for training tables).
  \item Define training and test sets for the training tables.
}
}
