% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcTest.R
\name{fcTest}
\alias{fcTest}
\title{Compute out-of-sample RMS Errors for model output}
\usage{
fcTest(input, output, subset = NULL)
}
\arguments{
\item{input}{Control object. The input control object used by \code{\link{fcRun}} to
generate the output.}

\item{output}{Model Output. The model output produced by \code{\link{fcRun}}.}

\item{subset}{The subset of the data over which to estimate RMS Errors.
I include this because in some cases the test subset has been different from
the training subset in non-random ways.}
}
\value{
This returns a list with the following members:

\describe{
  \item{lhs}{Name of the left-hand side variable.}
  \item{subset}{The subset to which the results are applied.}
  \item{se}{A named vector with the root-mean-square errors on the
            out-of-sample data for each model in the control object.}
  \item{results}{A data frame with the row-by-row results.}
}
}
\description{
Compute out-of-sample RMS Errors for model output
}
\details{
This function takes output from the \code{\link{fcRun}} function and calculates the
out-of-sample Root-Mean-Square Error values for each model in the output
object.
}
